requirements.txt
Flask
requests
beautifulsoup4
googlesearch-python
numpy
langchain
langchain-ollama
